Change type for total credits in organization

+++ Domain/Entities/OrganizationEntity.cs
@@ -15,7 +15,7 @@ public class OrganizationEntity : BaseEntity
     public string DisplayName { get; set; }
 
     [Required]
-    public int TotalCredits { get; private set; }
+    public decimal TotalCredits { get; private set; }
 
     [Required]
     public PartnershipEnum Partnership { get; private set; }
@@ -25,9 +25,8 @@ public class OrganizationEntity : BaseEntity
     
     [Required]
     public ICollection<CreditHistoryEntity>? CreditHistories{ get; set; }
-    // public IEnumerable<PurchaseEntity>? Purchases { get; set; }
-
-    public OrganizationEntity(Guid id, string alias, string displayName, int totalCredits, PartnershipEnum partnership, DateTime created, DateTime modified, ICollection<CreditHistoryEntity>? creditHistories) : base(id, created, modified)
+    
+    public OrganizationEntity(Guid id, string alias, string displayName, decimal totalCredits, PartnershipEnum partnership, DateTime created, DateTime modified, ICollection<CreditHistoryEntity>? creditHistories) : base(id, created, modified)
     {
         Alias = alias;
         DisplayName = displayName;
